Add DocumentHieratchicalInnerAttention and GPU/CPU test passed

+++ utils/masked_softmax.py
@@ -13,13 +13,14 @@ class MaskedSoftmax(nn.Module):
 
     def forward(self, input, mask=None):
         if mask is None:
-            mask = torch.ones_like(input).long()
+            mask = torch.ones_like(input).short()
         if mask.dim() == input.dim()-1:
             # in case of mask may not be padded in hidden feature space
             mask = mask.unsqueeze(-1).expand(input.shape)
         if input.is_cuda:
             mask = mask.cuda()
-
+        # print(mask.shape)
+        # print(input.shape)
         assert mask.shape == input.shape and mask.dtype == torch.short, "Mask shape %s should match input shape %s" % (mask.shape, input.shape)
         input.masked_fill_(mask == 0, -1e9)
         input = self.softmax(input)
